+++ src/services/Plugins.php
@@ -1151,7 +1151,8 @@ class Plugins extends Component
      */
     public function getPluginLicenseKey(string $handle)
     {
-        return $this->normalizePluginLicenseKey(App::parseEnv($this->getStoredPluginInfo($handle)['licenseKey'] ?? null));
+        $licenseKey = $this->getStoredPluginInfo($handle)['licenseKey'] ?? null;
+        return $this->normalizePluginLicenseKey(App::parseEnv($licenseKey) ?: $licenseKey);
     }
 
     /**