+++ src/services/Gc.php
@@ -10,6 +10,7 @@ namespace craft\services;
 use Craft;
 use craft\base\ElementInterface;
 use craft\config\GeneralConfig;
+use craft\console\Application as ConsoleApplication;
 use craft\db\Query;
 use craft\db\Table;
 use craft\elements\Asset;
@@ -130,7 +131,7 @@ class Gc extends Component
             return;
         }
 
-        Console::stdout("    > deleting trashed volumes and their folders ... ");
+        $this->_stdout("    > deleting trashed volumes and their folders ... ");
         $condition = $this->getHardDeleteConditions($generalConfig);
 
         $volumes = (new Query())->select(['id'])->from([Table::VOLUMES])->where($condition)->all();
@@ -150,7 +151,7 @@ class Gc extends Component
         }
 
         Volume::deleteAll(['id' => $volumeIds]);
-        Console::stdout("done\n", Console::FG_GREEN);
+        $this->_stdout("done\n", Console::FG_GREEN);
     }
 
     /**
@@ -172,9 +173,9 @@ class Gc extends Component
         }
 
         foreach ($tables as $table) {
-            Console::stdout("    > deleting trashed rows in the `$table` table ... ");
+            $this->_stdout("    > deleting trashed rows in the `$table` table ... ");
             Db::delete($table, $condition);
-            Console::stdout("done\n", Console::FG_GREEN);
+            $this->_stdout("done\n", Console::FG_GREEN);
         }
     }
 
@@ -214,7 +215,7 @@ SQL;
         }
 
         $db->createCommand($sql, ['type' => $elementType])->execute();
-        Console::stdout("done\n", Console::FG_GREEN);
+        $this->_stdout("done\n", Console::FG_GREEN);
     }
 
     private function _purgeUnsavedDrafts(GeneralConfig $generalConfig)
@@ -223,9 +224,9 @@ SQL;
             return;
         }
 
-        Console::stdout('    > purging unsaved drafts that have gone stale ... ');
+        $this->_stdout('    > purging unsaved drafts that have gone stale ... ');
         Craft::$app->getDrafts()->purgeUnsavedDrafts();
-        Console::stdout("done\n", Console::FG_GREEN);
+        $this->_stdout("done\n", Console::FG_GREEN);
     }
 
     private function _purgePendingUsers(GeneralConfig $generalConfig)
@@ -234,9 +235,9 @@ SQL;
             return;
         }
 
-        Console::stdout('    > purging pending users with stale activation codes ... ');
+        $this->_stdout('    > purging pending users with stale activation codes ... ');
         Craft::$app->getUsers()->purgeExpiredPendingUsers();
-        Console::stdout("done\n", Console::FG_GREEN);
+        $this->_stdout("done\n", Console::FG_GREEN);
     }
 
     /**
@@ -246,7 +247,7 @@ SQL;
      */
     public function removeEmptyTempFolders(): void
     {
-        Console::stdout('    > removing empty temp folders ... ');
+        $this->_stdout('    > removing empty temp folders ... ');
 
         $emptyFolders = (new Query())
             ->from(['folders' => Table::VOLUMEFOLDERS])
@@ -268,7 +269,7 @@ SQL;
         }
 
         VolumeFolder::deleteAll(['id' => array_keys($emptyFolders)]);
-        Console::stdout("done\n", Console::FG_GREEN);
+        $this->_stdout("done\n", Console::FG_GREEN);
     }
 
     /**
@@ -280,12 +281,12 @@ SQL;
             return;
         }
 
-        Console::stdout('    > deleting stale user sessions ... ');
+        $this->_stdout('    > deleting stale user sessions ... ');
         $interval = DateTimeHelper::secondsToInterval($generalConfig->purgeStaleUserSessionDuration);
         $expire = DateTimeHelper::currentUTCDateTime();
         $pastTime = $expire->sub($interval);
         Db::delete(Table::SESSIONS, ['<', 'dateUpdated', Db::prepareDateForDb($pastTime)]);
-        Console::stdout("done\n", Console::FG_GREEN);
+        $this->_stdout("done\n", Console::FG_GREEN);
     }
 
     /**
@@ -293,9 +294,9 @@ SQL;
      */
     private function _deleteStaleAnnouncements(): void
     {
-        Console::stdout('    > deleting stale feature announcements ... ');
+        $this->_stdout('    > deleting stale feature announcements ... ');
         Db::delete(Table::ANNOUNCEMENTS, ['<', 'dateRead', Db::prepareDateForDb(new DateTime('7 days ago'))]);
-        Console::stdout("done\n", Console::FG_GREEN);
+        $this->_stdout("done\n", Console::FG_GREEN);
     }
 
 
@@ -304,7 +305,7 @@ SQL;
      */
     private function _deleteOrphanedDraftsAndRevisions(): void
     {
-        Console::stdout('    > deleting orphaned drafts and revisions ... ');
+        $this->_stdout('    > deleting orphaned drafts and revisions ... ');
         $db = Craft::$app->getDb();
         $elementsTable = Table::ELEMENTS;
 
@@ -329,14 +330,14 @@ SQL;
             $db->createCommand($sql)->execute();
         }
 
-        Console::stdout("done\n", Console::FG_GREEN);
+        $this->_stdout("done\n", Console::FG_GREEN);
     }
 
     private function _deleteOrphanedSearchIndexes(): void
     {
-        Console::stdout('    > deleting orphaned search indexes ... ');
+        $this->_stdout('    > deleting orphaned search indexes ... ');
         Craft::$app->getSearch()->deleteOrphanedIndexes();
-        Console::stdout("done\n", Console::FG_GREEN);
+        $this->_stdout("done\n", Console::FG_GREEN);
     }
 
     private function _gcCache(): void
@@ -364,7 +365,7 @@ SQL;
             return;
         }
 
-        Console::stdout('    > garbage-collecting data caches ... ');
+        $this->_stdout('    > garbage-collecting data caches ... ');
 
         if ($hasForceArg) {
             $cache->gc(true);
@@ -372,7 +373,7 @@ SQL;
             $cache->gc();
         }
 
-        Console::stdout("done\n", Console::FG_GREEN);
+        $this->_stdout("done\n", Console::FG_GREEN);
     }
 
     /**
@@ -394,5 +395,12 @@ SQL;
             ];
         }
         return $condition;
+    }
+
+    private function _stdout(string $string, ...$format): void
+    {
+        if (Craft::$app instanceof ConsoleApplication) {
+            Console::stdout($string, ...$format);
+        }
     }
 }