+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Fixed
 - Fixed a bug where the “Edit files uploaded by other users” volume permission was useless unless the “Upload files” permission had been granted. ([#11818](https://github.com/craftcms/cms/issues/11818))
+- Fixed a bug where the “Save and add another” entry action was being shown for users without the “Create entries” section permission. ([#11819](https://github.com/craftcms/cms/issues/11819))
 
 ## 3.7.52 - 2022-08-23