+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Release Notes for Craft CMS 4
 
+## Unreleased
+
+### Fixed
+- Fixed a bug where Craft was overwriting plugin license keys set to an environment variable, if the environment variable didn’t have a value. ([#11830](https://github.com/craftcms/cms/issues/11830))
+
 ## 4.2.2 - 2022-08-23
 
 ### Added