+++ CHANGELOG.md
@@ -2,8 +2,8 @@
 
 ## Unreleased
 
-### Fixed
-- Fixed a bug where Craft was overwriting plugin license keys set to an environment variable, if the environment variable didn’t have a value. ([#11830](https://github.com/craftcms/cms/issues/11830))
+### Changed
+- If a plugin’s license key is set to an empty environment variable, its trial license key will now be stored in `.env` rather than the project config. ([#11830](https://github.com/craftcms/cms/issues/11830))
 
 ## 4.2.2 - 2022-08-23